fix: update reset.sh to accept osm-pbf file as argument

+++ reset.sh
@@ -2,10 +2,14 @@
 
 set -euo pipefail
 
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 <osm-pbf-file>"
+    exit 1
+fi
 
 docker compose down
 sudo rm -rf /osm-data/*
-sudo cp ~/Downloads/chad-latest.osm.pbf ./data.osm.pbf
+sudo cp "$1" ./data.osm.pbf
 docker compose build
 docker compose up import
 docker compose up light dark
